fully replace drone connection status
