+++ radio/app/controllers/armController.py
@@ -4,7 +4,7 @@ import time
 from threading import current_thread
 from typing import TYPE_CHECKING
 
-from app.signals import drone_error
+from app.signals import DroneError
 from app.customTypes import Response
 from app.utils import commandAccepted, sendingCommandLock
 from pymavlink import mavutil
@@ -90,8 +90,8 @@ class ArmController:
             self.drone.logger.error(e, exc_info=True)
             self.drone.error(str(e))
 
-            ## Or you can have:                 (TODO: remove)
-            drone_error.send(str(e))
+            ## Or you can have just have:                 (TODO: remove)
+            DroneError.send(str(e))
 
             return {
                 "success": False,
